+++ lib/l10n/app_localizations_de.dart
@@ -90,6 +90,12 @@ class AppLocalizationsDe extends AppLocalizations {
   String get libraryAll => 'Alle';
 
   @override
+  String get libraryViewAll => 'All';
+
+  @override
+  String get libraryViewOverview => 'Overview';
+
+  @override
   String get libraryUncategorized => 'Ohne Kategorie';
 
   @override