+++ lib/l10n/app_localizations_pt.dart
@@ -90,6 +90,12 @@ class AppLocalizationsPt extends AppLocalizations {
   String get libraryAll => 'Tudo';
 
   @override
+  String get libraryViewAll => 'All';
+
+  @override
+  String get libraryViewOverview => 'Overview';
+
+  @override
   String get libraryUncategorized => 'Sem categoria';
 
   @override