+++ test/library_drawer_test.dart
@@ -23,6 +23,7 @@ void main() {
         home: LibraryDrawer(
           folders: folders,
           currentFolderId: null,
+          overviewActive: true,
           trashCount: 3,
           onOverview: () {},
           onOpenTrash: () {},